Разбил файл main.js на модули
